手動予約とルール予約の option の形式を同じに変更
