+++ src/server/Model/IPC/IPCServer.ts
@@ -3,10 +3,10 @@ import * as apid from '../../../../node_modules/mirakurun/api';
 import * as events from '../../IoEvents';
 import Model from '../Model';
 import { MirakurunManageModelInterface } from '../Operator/EPGUpdate/MirakurunManageModel';
-import { AddReserveInterface } from '../Operator/ManualReserveInterface';
 import { RecordedManageModelInterface } from '../Operator/Recorded/RecordedManageModel';
 import { RecordingManageModelInterface } from '../Operator/Recording/RecordingManageModel';
 import { ReservationManageModelInterface } from '../Operator/Reservation/ReservationManageModel';
+import { AddReserveInterface } from '../Operator/ReserveProgramInterface';
 import { RuleManageModelInterface } from '../Operator/Rule/RuleManageModel';
 import { RuleInterface } from '../Operator/RuleInterface';
 import { EncodeProgram } from '../Service/Encode/EncodeManageModel';
