+++ src/server/Model/Operator/ReserveProgramInterface.ts
@@ -1,12 +1,13 @@
+import * as apid from '../../../../node_modules/mirakurun/api';
 import * as DBSchema from '../DB/DBSchema';
-import { AddReserveOptionInterface } from './ManualReserveInterface';
-import { EncodeInterface, OptionInterface } from './RuleInterface';
+import { EncodeInterface } from './RuleInterface';
 
 export interface ReserveProgram {
     program: DBSchema.ProgramSchema;
     encodeOption?: EncodeInterface;
     isSkip: boolean;
     isConflict: boolean;
+    option?: ReserveOptionInterface;
 }
 
 /**
@@ -14,14 +15,26 @@ export interface ReserveProgram {
  */
 export interface RuleReserveProgram extends ReserveProgram {
     ruleId?: number;
-    ruleOption?: OptionInterface;
 }
 
 /**
  * 手動予約
  */
 export interface ManualReserveProgram extends ReserveProgram {
-    manualOption?: AddReserveOptionInterface;
     manualId?: number;
+}
+
+export interface ReserveOptionInterface {
+    directory?: string;
+    recordedFormat?: string;
+}
+
+/**
+ * 予約追加時データ
+ */
+export interface AddReserveInterface {
+    programId: apid.ProgramId;
+    option?: ReserveOptionInterface;
+    encode?: EncodeInterface;
 }
